Move CAS1 App Creation Logic into CAS1 Service
This commit introduces a new `Cas1ApplicationCreationService` that is responsible for all logic related to creating, updating and submitting a new application. This is in preparation for some fixes around status logic
